Added 'AccesedValue' to CheckResult enum and enhanced analyzer checks.
'AccesedValue' enum type was added to WalkerResult to provide handling of more cases in the ResultValueWalker logic. This will improve the analyzer's functionality in catching incorrect usage of 'result.Value' without an appropriate 'IsSuccess' or 'IsError' check. The modification aims to improve the diagnostic output of the analyzers to minimize possible errors. Additional changes were made in 'UseResultValueWithoutCheck_older.cs' and 'ResultValueWalker.cs' to incorporate this new enum and improve result checking.